requirements.txt
flask~=2.0.1
werkzeug~=2.0.3
google-generativeai~=0.3.0
python-dotenv~=1.0.0
PyPDF2~=3.0.1
Pillow~=10.0.0
requests~=2.31.0
flask-cors~=5.0.1
python-dotenv~=1.0.0
